+++ internal/cache/user_cache.go
@@ -10,9 +10,9 @@ import (
 	"strings"
 
 	// External packages
-	cache "github.com/soulteary/cache-kit"
+	cache "github.com/soulteary/cache-kit/v2"
 	"github.com/soulteary/cli-kit/validator"
-	secure "github.com/soulteary/secure-kit"
+	secure "github.com/soulteary/secure-kit/v2"
 
 	// Internal packages
 	"github.com/soulteary/warden/internal/define"