+++ internal/tracing/middleware.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"net/http"
 
-	common_tracing "github.com/soulteary/tracing-kit"
+	common_tracing "github.com/soulteary/tracing-kit/v2"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"